repair bulletin board

+++ app/Http/Controllers/Backend/Admin/BulletinBoardsController.php
@@ -292,22 +292,23 @@ class BulletinBoardsController extends Controller
                     }else{
                         $bulletin_board->author = $request->author;
                     }
-
+dd(Input::hasFile('image'));
                     if($request->hasFile('image')) {
                         if($request->action == 'update'){                        
                             if($bulletin_board->img_url != ""){  
-                            $image_path = public_path().'/storage/news/'.$bulletin_board->img_url;
-                            unlink($image_path);
+                                $image_path = public_path().'/storage/news/'.$bulletin_board->img_url;
+                                if(file_exists($image_path))
+                                    unlink($image_path);
                             }
                         }
                         createdirYmd('storage');
+                        createdirYmd('storage/news');
                         $file = Input::file('image');            
                         $name = str_random(20). '-' .$file->getClientOriginalName();  
                         $bulletin_board->img_url = date("Y")."/".date("m")."/".date("d")."/".$name;          
-                        // $file->move(public_path().'/storage/news/'.date("Y")."/".date("m")."/".date("d")."/", $name);
-                        createdirYmd('storage/news');
-                        $path = public_path('/storage/news/'.date("Y")."/".date("m")."/".date("d")."/". $name);
-                        resizeAndSaveImage($file, $path);
+                        $file->move(public_path().'/storage/news/'.date("Y")."/".date("m")."/".date("d")."/", $name);
+                        // $path = public_path('/storage/news/'.date("Y")."/".date("m")."/".date("d")."/". $name);
+                        // resizeAndSaveImage($file, $path);
                     }
               
                     $bulletin_board->save();
